Moved the allOf discriminator tests into their own test class
